Splitting StyxHttpClient functionality into two separate logically types. It is a first step in a refactoring process of decoupling the logic.
