+++ components/api/src/main/java/com/hotels/styx/api/HttpClient.java
@@ -17,12 +17,10 @@ package com.hotels.styx.api;
 
 import rx.Observable;
 
-import java.io.Closeable;
-
 /**
  * HTTP Client that returns an observable of response.
  */
-public interface HttpClient extends Closeable {
+public interface HttpClient {
     /**
      * Processes a new request.
      *
@@ -34,10 +32,4 @@ public interface HttpClient extends Closeable {
      *
      */
     Observable<HttpResponse> sendRequest(HttpRequest request);
-
-    default void close() {
-    }
-
-    default void registerStatusGauges() {
-    }
 }
