+++ components/proxy/src/main/java/com/hotels/styx/routing/handlers/BackendServiceProxy.java
@@ -27,9 +27,11 @@ import com.hotels.styx.infrastructure.Registry;
 import com.hotels.styx.infrastructure.configuration.yaml.JsonNodeConfig;
 import com.hotels.styx.proxy.BackendServiceClientFactory;
 import com.hotels.styx.proxy.BackendServicesRouter;
+import com.hotels.styx.proxy.ConnectionPoolProviderFactory;
 import com.hotels.styx.proxy.ProxyServerConfig;
 import com.hotels.styx.proxy.RouteHandlerAdapter;
 import com.hotels.styx.proxy.StyxBackendServiceClientFactory;
+import com.hotels.styx.proxy.StyxConnectionPoolProviderFactory;
 import com.hotels.styx.routing.config.BuiltinHandlersFactory;
 import com.hotels.styx.routing.config.HttpHandlerFactory;
 import com.hotels.styx.routing.config.RoutingConfigDefinition;
@@ -50,8 +52,10 @@ public class BackendServiceProxy implements HttpHandler2 {
 
     private final RouteHandlerAdapter handler;
 
-    private BackendServiceProxy(BackendServiceClientFactory serviceClientFactory, Registry<BackendService> registry) {
-        BackendServicesRouter router = new BackendServicesRouter(serviceClientFactory);
+    private BackendServiceProxy(BackendServiceClientFactory serviceClientFactory,
+                                ConnectionPoolProviderFactory connectionPoolProviderFactory,
+                                Registry<BackendService> registry) {
+        BackendServicesRouter router = new BackendServicesRouter(serviceClientFactory, connectionPoolProviderFactory);
         registry.addListener(router);
         handler = new RouteHandlerAdapter(router);
     }
@@ -67,21 +71,29 @@ public class BackendServiceProxy implements HttpHandler2 {
     public static class ConfigFactory implements HttpHandlerFactory {
         private final BackendServiceClientFactory serviceClientFactory;
         private final Map<String, Service> services;
+        private final ConnectionPoolProviderFactory connectionPoolProviderFactory;
 
         private static StyxBackendServiceClientFactory serviceClientFactory(Environment environment) {
+            return new StyxBackendServiceClientFactory(environment);
+        }
+
+        private static ConnectionPoolProviderFactory connectionPoolProviderFactory(Environment environment) {
             ProxyServerConfig proxyConfig = environment.styxConfig().proxyServerConfig();
-            return new StyxBackendServiceClientFactory(environment, proxyConfig.clientWorkerThreadsCount());
+            return new StyxConnectionPoolProviderFactory(environment, proxyConfig.clientWorkerThreadsCount());
         }
 
         @VisibleForTesting
-        ConfigFactory(BackendServiceClientFactory serviceClientFactory, Map<String, Service> services) {
+        ConfigFactory(BackendServiceClientFactory serviceClientFactory, ConnectionPoolProviderFactory connectionPoolProviderFactory,
+                      Map<String, Service> services) {
             this.serviceClientFactory = serviceClientFactory;
+            this.connectionPoolProviderFactory = connectionPoolProviderFactory;
             this.services = services;
         }
 
         public ConfigFactory(Environment environment, Map<String, Service> services) {
             this.services = services;
             this.serviceClientFactory = serviceClientFactory(environment);
+            this.connectionPoolProviderFactory = connectionPoolProviderFactory(environment);
         }
 
         @Override
@@ -103,7 +115,7 @@ public class BackendServiceProxy implements HttpHandler2 {
             }
             Registry<BackendService> registry = (Registry<BackendService>) service;
 
-            return new BackendServiceProxy(serviceClientFactory, registry);
+            return new BackendServiceProxy(serviceClientFactory, connectionPoolProviderFactory, registry);
         }
     }
